// src/app/Models/PersonalProposal/guarantor-income.ts
export class GuarantorIncome {

    ID: number;
    G_ID: number
    PROPOSAL_ID: number;

    IS_SALARY: boolean = false;
    IS_BUSINESS: boolean = false;
    IS_AGRICULTURE: boolean = false;
    IS_HOUSE: boolean = false;
    IS_OTHER: boolean = false;

    IS_EXPENDITURE: boolean = false;
    IS_NO_INCOME: boolean = false;

}